Silent warnings about ignored return value of write
The image.c does never check the return value of the 'write' function. This results
in many compiler warnings (only the first one quoted here):

image.c: In function 'ImageWriteGif':
image.c:402:3: warning: ignoring return value of 'write', declared with attribute warn_unused_result [-Wunused-result]
   write(fh, buf, n);
   ^

Adding a check for the call is not trivial, as the enclosing functions do not really
provide a chance to return with an error. But changing these functions will break API/ABI.

So for the moment, it's simpler to silent the warning by explicite suppressing it
for the relevant file. After a proper fix is worked out, this commit should be reverted.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>